+++ R/conditionalEffectPlot.R
@@ -1,16 +1,28 @@
 #'Make conditional effect plot
-#'@param fit An object of class lavaan
+#'@param semfit An object of class lavaan
 #'@param values Optional value
 #'@param data A data.frame
 #'@param mod Name of moderator variable
+#'@param color charactor vector line color
+#'@param lty numeric line type
+#'@param linesize numeric linesize
 #'@importFrom ggplot2 stat_function guides xlab ylab ggplot aes_string coord_fixed
 #'@importFrom ggplot2 geom_text aes theme_bw
+#'@importFrom predict3d theme_bw2
 #'@export
-conditionalEffectPlot=function(fit,values=NULL,data,mod="skeptic"){
-    res=parameterEstimates(fit)
+conditionalEffectPlot=function(semfit,values=NULL,data,
+                               mod=NULL,color=c("black","red"),lty=c(1,3),linesize=1){
+
+    # values=NULL;mod=NULL;color=c("black","red");lty=c(1,3);linesize=1
+
+    res=parameterEstimates(semfit)
     res=res[res$label!="",]
     res
-    x=modmedSummary(fit,mod=mod,values=values)
+    if(is.null(mod)){
+        mod=res$lhs[str_detect(res$label,"mean")][1]
+        mod
+    }
+    x=modmedSummary(semfit,mod=mod,values=values)
     indirect=attr(x,"indirect")
     direct=attr(x,"direct")
     indirect
@@ -22,8 +34,8 @@ conditionalEffectPlot=function(fit,values=NULL,data,mod="skeptic"){
     fun1=function(W) eval(parse(text=direct))
     fun2=function(W) eval(parse(text=indirect))
     p<-ggplot(data=data,aes_string(x=mod))+
-        stat_function(fun=fun1,color="black",size=1)+
-        stat_function(fun=fun2,color="red",lty=3,size=1)
+        stat_function(fun=fun1,color=color[1],lty=lty[1],size=linesize)+
+        stat_function(fun=fun2,color=color[2],lty=lty[2],size=linesize)
     p
     res1=fun2eq(fun1)
     res2=fun2eq(fun2)
@@ -39,8 +51,11 @@ conditionalEffectPlot=function(fit,values=NULL,data,mod="skeptic"){
     df$slope2=df$slope*ratio
     df$radian=atan(df$slope2)
     df$angle=df$radian*180/pi
-    df$label=paste0("Y = ",sprintf("%0.3f",df$intercept),ifelse(df$slope>0," + "," - "),
-                    sprintf("%0.3f",abs(df$slope)),"*W")
+    df
+    # str(df)
+    df$label=paste0("Y = ",sprintf("%0.3f",df$intercept),
+                    ifelse(df$slope==0,"",ifelse(df$slope>0," + "," - ")),
+                    ifelse(df$slope==0,"",paste0(sprintf("%0.3f",abs(df$slope)),"*W")))
     p
     x=info$xmin+(info$xmax-info$xmin)/3*(1:2)
     x
@@ -52,11 +67,11 @@ conditionalEffectPlot=function(fit,values=NULL,data,mod="skeptic"){
     df$y=y
     df
     p+coord_fixed(ratio=ratio)+
-        geom_text(data=df,aes_string(x="x",y="y",label="name",angle="angle"),color=c("black","red"),vjust=-1.0,fontface="italic")+
-        geom_text(data=df,aes_string(x="x",y="y",label="label",angle="angle"),color=c("black","red"),fontface="italic",vjust=1.5)+
+        geom_text(data=df,aes_string(x="x",y="y",label="name",angle="angle"),color=color,vjust=-1.0,fontface="italic")+
+        geom_text(data=df,aes_string(x="x",y="y",label="label",angle="angle"),color=color,fontface="italic",vjust=1.5)+
         guides(color=FALSE)+
         xlab(paste0(mod,"(W)"))+ylab("Conditional Effects")+
-        theme_bw()
+        theme_bw2()
 }
 
 #'Make equation from function